Android doesn't support long double, so use sqrt instead of sqrtl. According to http://en.cppreference.com/w/c/numeric/math/sqrt, sqrt() is defined as a type-generic macro.

+++ modules/photometric_calib/src/ResponseCalib.cpp
@@ -41,7 +41,8 @@ Vec2d ResponseCalib::rmse(const double *G, const double *E, const std::vector<do
         }
     }
 
-    return Vec2d((double)(1e5 * sqrtl((e/num))), (double)num);
+    //return Vec2d((double)(1e5 * sqrtl((e/num))), (double)num);
+    return Vec2d((double)(1e5 *sqrt((e/num))), (double)num);
 }
 
 void ResponseCalib::plotE(const double* E, int w, int h, const std::string &saveTo)
